[feature] ability to definte the NTP Servers you want to use
previously, this container shipped with a static ntpd config file.
the most common reason people forked this repo was to simply update
the ntp server(s). as a result i have added the ability to pass
an `NTP_SERVER` list environment variable that is evaluated on
startup and the ntpd.conf file is dynamically changed accordingly.

this change also moves the default time server from google
(time[1-4].google.com) to time.cloudflare.com.