[feature] /!\ now running as non-root user /!\
this is something i've been meaning to do for a while... we all know
that it's not the best security practices to run processes in a
container as root. this update configured chrony to drop priviledges
after startup (-u user).

after you launch the container, you can now see the chronyd process
(pid 1) running as the chrony user:

```
$> docker exec -ti ntp ps
PID   USER     TIME  COMMAND
  1 chrony    0:00 /usr/sbin/chronyd -u chrony -d -x
619 root      0:00 ps
```

ref: https://chrony.tuxfamily.org/doc/4.0/chronyd.html