adding noclientlog chronyd option and minor cleanup
as requested by @Gontier-Julien, i have added the ability to set
the noclientlog config option via an environment variable.

additionally, the rtcsync option is being removed since this
container has no access to the system clock so it's not needed.

finally, i addressed a small bug in the local timeserver check.
it previously would have matches on any IP address that contained
127 in the 1-3 octets. which is not what we want... it needs to
match 127.0/8